More warning removal.
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@1014 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ sandbox/rick/warnings/kernel/classes/NumberStringClass.hpp
@@ -137,7 +137,7 @@ long number_create_integer(UCHAR *, long, int, int);
     RexxInteger *strictLessOrEqual(RexxObject *);
     RexxObject  *hashCode();
 
-    int         ULong(ULONG *);
+    int         ULong(size_t *);
     RexxNumberString *clone();
     void        setString(RexxString *);
     void        roundUp(ULONG);
